CrimeDataBase-main/requirements.txt
streamlit
mysql-connector-python
pandas
pydeck
numpy
matplotlib
geopy
plotly
